lib/requirements.txt
pyannote.audio>=3.1
torch>=2.0
